Feat : "프로토콜 추가에 따른 기능 추가 , 주위 섹터 얻는 함수 수정"
GameServer
- Patrol 프로토콜 추가에 따른 수정
 - ObjectStateChange 처리 수정
   - 기존에는 몬스터의 모든 상태 변경은 서버에서 담당했엇기 때문에 ObjectStateChange는 플레이어를 대상으로만 처리 했었음
   - 몬스터의 Patrol상태 추가로 인해 몬스터가 정찰 위치에 도착하면, Idle 상태 변경을 요청하도록 기능을 추가해서 몬스터를 대상으로도 ObjectStateChange를 처리해줘야함
   - 따라서 ObjectStateChange를 클라에서 요청할때 ObjectType과 ObjectId를 받아서 대상을 찾고 상태 값 수정하도록 변경
 - 정찰 패킷 조합 함수 추가

Channel
- 주위 섹터 얻는 함수 수정
 - 매개변수로 받는 Range의 의미를 거리값이 아닌 주위 섹터 거리를 의미하도록 변경
 - 기존에는 거리값을 받고 좌표로부터 기준 거리 왼쪽위, 오른쪽 아래 좌표를 구하고 구해준 좌표를 토대로 섹터를 구했으나,
   수정한 방식은 Range가 섹터거리를 의미하므로 좌표로 부터 섹터를 구하고 해당 섹터를 기준으로 Range 위치에 존재하는 섹터를 반환해주도록 변경
 - 따라서 주위 섹터를 얻을때 섹터 크기인 '10'을 전달해줬으나 이제부터는 '1'을 전달해주면 현재 섹터 기준 '1'거리만큼 있는 섹터들을 구해서 반환해줌

GameObjectInfo
- 정찰 상태 값 추가

GameObjectInfo
- 정찰 상태 브로드캐스팅 추가

+++ GameServer/Map.h
@@ -53,9 +53,9 @@ struct st_Vector2Int
 	}
 	
 	// 거리 구하기
-	static int Distance(st_Vector2Int TargetCellPosition, st_Vector2Int MyCellPosition)
+	static int16 Distance(st_Vector2Int TargetCellPosition, st_Vector2Int MyCellPosition)
 	{
-		return (int)sqrt(pow(TargetCellPosition._X - MyCellPosition._X, 2) + pow(TargetCellPosition._Y - MyCellPosition._Y, 2));
+		return (int16)sqrt(pow(TargetCellPosition._X - MyCellPosition._X, 2) + pow(TargetCellPosition._Y - MyCellPosition._Y, 2));
 	}
 
 	//--------------------------------------------------------------
